requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.4
google-generativeai==0.7.2
python-dotenv==1.0.1
